ALLJOYN-2210 Fixed compiler warnings from Windows VC++
ICECandidatePair was reporting a warning C4244 'return' conversion from double
to uint32_t possible loss of data. The double was being returned from the
GetQueuedTimeOffset() method from StunRetry.

I talked with Padama and she verified that it was safe to change the return
value of the GetQueuedTimeOffset function.

Fixed a few implicit other casting warnings that could result in possible loss of data.

Change-Id: Id0f71dc5cd73eab353659d057b1dd34030df650d