ALLJOYN-2324 fix issue in ObjSecurityTest that would report false failures
In several of the tests it was should have been waiting for a signal to be
received.  It was waiting for the object to be registered but not on the
signal. This could cause it to check the value of the signal before it had
a chance to be received.

Change-Id: If50bd7c04470bb157ed07cd5e5d02c93dd7f896b