ALLJOYN-2441 Doxygen builder will once again produce PDF output
For a long while the doxygen builder has failed when producing PDF output
the reason for this failure was primarily error in the quic.sty file that added
our custom footer to each page of the document.

Our use of the \renewcommand was also done improperly and need to be removed.

Added the LATEX_FOOTER and LATEX_EXTRA_FILES flag to the doxygen config file.

Change-Id: I191e0a2dd42a5178760cddeda64ebf86e49df2c5